Moved auth token processing code into library
I did this to properly separate concerns for processing new auth tokens. The
user object had far too much knowledge of auth tokens.
